Move attachments API related code to a separate file
