+++ frontend/src/lib-common/api-types/attachment.ts
@@ -1,0 +1,7 @@
+import { UUID } from '../types'
+
+export interface Attachment {
+  id: UUID
+  name: string
+  contentType: string
+}
