+++ frontend/src/lib-components/molecules/FileDownloadButton.tsx
@@ -7,7 +7,7 @@ import styled from 'styled-components'
 
 import { Result } from 'lib-common/api'
 import { downloadBlobAsFile } from 'lib-common/utils/file'
-import { Attachment } from 'lib-common/api-types/application/ApplicationDetails'
+import { Attachment } from 'lib-common/api-types/attachment'
 import { UUID } from 'lib-common/types'
 
 const DownloadButton = styled.button`
